Update :Remove the unecessary service and update on mono and clean arch structure
