+++ backend/.env.example
@@ -1,0 +1,24 @@
+# Auth Service Environment Variables
+
+# Database Configuration
+DB_HOST=localhost
+DB_PORT=3306
+DB_USER=root
+DB_PASSWORD=trade%2025
+DB_NAME=password
+LOG_LEVEL=debug
+
+# JWT Configuration
+JWT_SECRET="TTBOX_VOLTIX_22DAJJJHAJHSHDJSADJHSAD"
+JWT_EXPIRATION_HOURS=72
+
+# Google OAuth Configuration
+GOOGLE_CLIENT_ID=your-google-client-id.apps.googleusercontent.com
+GOOGLE_CLIENT_SECRET=your-google-client-secret
+GOOGLE_REDIRECT_URL=http://localhost:3001/api/auth/google/callback
+
+# CORS
+CORS_ORIGINS=http://localhost:5173,http://localhost:3000
+
+# Server
+PORT=3001
