+++ backend/cmd/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/ratheeshkumar25/Voltix-Trades-Bot/internal/di"
+
+func main() {
+	di.Init()
+}
