+++ backend/internal/db/db.go
@@ -1,0 +1,49 @@
+package db
+
+import (
+	"fmt"
+	"log"
+
+	"github.com/ratheeshkumar25/Voltix-Trades-Bot/internal/config"
+	models "github.com/ratheeshkumar25/Voltix-Trades-Bot/internal/modles"
+	"gorm.io/driver/mysql"
+	"gorm.io/gorm"
+)
+
+// DB is the global database connection
+var DB *gorm.DB
+
+func ConnectDB(config *config.Config) *gorm.DB {
+	host := config.DBHost
+	port := config.DBPort
+	user := config.DBUser
+	password := config.DBPassword
+	dbname := config.DBName
+	sslmode := config.DBSSLMode
+
+	//
+	log.Printf("Connecting to DB: host=%s, user=%s, password=%s, dbname=%s, port=%s, sslmode=%s\n", host, user, password, dbname, port, sslmode)
+
+	//construct the DSN connection correctly
+	dsn := fmt.Sprintf("host=%s user=%s password=%s dbname=%s port=%s sslmode=%s TimeZone=UTC", host, user, password, dbname, port, sslmode)
+
+	db, err := gorm.Open(mysql.Open(dsn), &gorm.Config{})
+	if err != nil {
+		log.Fatalf("Failed to connect to database: %v", err)
+	}
+
+	err = db.AutoMigrate(
+		&models.User{},
+		&models.Subscription{},
+		&models.Notification{},
+		&models.Session{},
+	)
+
+	if err != nil {
+		log.Printf("error while migrating %v", err.Error())
+		return nil
+	}
+	DB = db // Assign to global variable
+	return db
+
+}
