+++ backend/internal/di/di.go
@@ -1,0 +1,10 @@
+package di
+
+import "github.com/ratheeshkumar25/Voltix-Trades-Bot/internal/config"
+
+func Init() {
+	config, err := config.LoadConfig()
+	if err != nil {
+		panic(err)
+	}
+}
