--- backend/services/auth-service/pkg/oauth/jwt.go
@@ -1,101 +1,0 @@
-package oauth
-
-import (
-	"crypto/rand"
-	"encoding/base64"
-	"errors"
-	"ratheeshkumar25/github.com/trading_bot/auth-service/internal/models"
-	"time"
-
-	"github.com/golang-jwt/jwt/v5"
-	"github.com/google/uuid"
-	"golang.org/x/crypto/bcrypt"
-)
-
-var (
-	jwtSecret       = []byte(getJWTSecret())
-	ErrInvalidToken = errors.New("invalid token")
-)
-
-// JWTClaims represents JWT token claims
-type JWTClaims struct {
-	UserID uuid.UUID `json:"user_id"`
-	Email  string    `json:"email"`
-	Role   string    `json:"role"`
-	jwt.RegisteredClaims
-}
-
-// GenerateJWT creates a new JWT token for a user
-func GenerateJWT(user *models.User) (string, error) {
-	claims := JWTClaims{
-		UserID: user.ID,
-		Email:  user.Email,
-		Role:   string(user.Role),
-		RegisteredClaims: jwt.RegisteredClaims{
-			ExpiresAt: jwt.NewNumericDate(time.Now().Add(24 * time.Hour)),
-			IssuedAt:  jwt.NewNumericDate(time.Now()),
-			NotBefore: jwt.NewNumericDate(time.Now()),
-			Issuer:    "voltix-trading-bot",
-			Subject:   user.ID.String(),
-		},
-	}
-
-	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
-	return token.SignedString(jwtSecret)
-}
-
-// ValidateJWT validates and parses a JWT token
-func ValidateJWT(tokenString string) (*JWTClaims, error) {
-	token, err := jwt.ParseWithClaims(tokenString, &JWTClaims{}, func(token *jwt.Token) (interface{}, error) {
-		if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
-			return nil, errors.New("unexpected signing method")
-		}
-		return jwtSecret, nil
-	})
-
-	if err != nil {
-		return nil, err
-	}
-
-	if claims, ok := token.Claims.(*JWTClaims); ok && token.Valid {
-		return claims, nil
-	}
-
-	return nil, ErrInvalidToken
-}
-
-// HashPassword hashes a password using bcrypt
-func HashPassword(password string) (string, error) {
-	bytes, err := bcrypt.GenerateFromPassword([]byte(password), 14)
-	return string(bytes), err
-}
-
-// CheckPassword compares password hash with plain password
-func CheckPassword(password, hash string) bool {
-	err := bcrypt.CompareHashAndPassword([]byte(hash), []byte(password))
-	return err == nil
-}
-
-// GenerateSecureToken generates a cryptographically secure random token
-func GenerateSecureToken(length int) (string, error) {
-	bytes := make([]byte, length)
-	if _, err := rand.Read(bytes); err != nil {
-		return "", err
-	}
-	return base64.URLEncoding.EncodeToString(bytes), nil
-}
-
-// getJWTSecret retrieves JWT secret from environment or generates one
-func getJWTSecret() string {
-	// In production, this MUST come from environment variable
-	if secret := getEnv("JWT_SECRET", ""); secret != "" {
-		return secret
-	}
-	// Fallback for development only - DO NOT use in production
-	return "dev-secret-change-in-production-min-32-chars-long"
-}
-
-func getEnv(key, fallback string) string {
-	// Simple env reader - would import from config in production
-	return fallback
-}
