--- backend/services/profit-prediction-service/cmd/main.go
@@ -1,31 +1,0 @@
-package main
-
-import (
-	"log"
-	"os"
-	"os/signal"
-	"syscall"
-
-	"github.com/nats-io/nats.go"
-)
-
-func main() {
-	log.Println("Starting Profit Prediction Service...")
-
-	// Connect to NATS
-	nc, err := nats.Connect(nats.DefaultURL)
-	if err != nil {
-		log.Fatalf("Failed to connect to NATS: %v", err)
-	}
-	defer nc.Close()
-
-	// Subscribe to market data updates? Or publish predictions?
-	// For now just a placeholder
-
-	log.Println("Profit Prediction Service is running")
-
-	// Wait for interrupt signal
-	c := make(chan os.Signal, 1)
-	signal.Notify(c, os.Interrupt, syscall.SIGTERM)
-	<-c
-}
