--- backend/services/profit-prediction-service/go.mod
@@ -1,13 +1,0 @@
-module ratheeshkumar25/github.com/trading_bot/profit-prediction-service
-
-go 1.25.4
-
-require github.com/nats-io/nats.go v1.47.0
-
-require (
-	github.com/klauspost/compress v1.18.0 // indirect
-	github.com/nats-io/nkeys v0.4.11 // indirect
-	github.com/nats-io/nuid v1.0.1 // indirect
-	golang.org/x/crypto v0.37.0 // indirect
-	golang.org/x/sys v0.32.0 // indirect
-)
